# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR & Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  admin: "admin",
  hr_manager: "hr_manager",
  employee: "employee",
} as const;

export interface AuthUser {
  id: number;
  username: string;
  email: string;
  role: AuthUserRole;
  lastLogin?: string | null;
}

export interface AuthResponse {
  user: AuthUser;
  token: string;
  lastLogin?: string | null;
}

export interface Department {
  id: number;
  name: string;
  description?: string | null;
  managerId?: number | null;
  employeeCount: number;
  createdAt: string;
}

export interface CreateDepartmentBody {
  name: string;
  description?: string | null;
  managerId?: number | null;
}

export type EmployeeStatus =
  (typeof EmployeeStatus)[keyof typeof EmployeeStatus];

export const EmployeeStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  position: string;
  departmentId?: number | null;
  departmentName?: string | null;
  hireDate: string;
  salary: number;
  status: EmployeeStatus;
  photoUrl?: string | null;
  address?: string | null;
  createdAt: string;
}

export type CreateEmployeeBodyStatus =
  (typeof CreateEmployeeBodyStatus)[keyof typeof CreateEmployeeBodyStatus];

export const CreateEmployeeBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateEmployeeBody {
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  position: string;
  departmentId?: number | null;
  hireDate: string;
  salary: number;
  status: CreateEmployeeBodyStatus;
  address?: string | null;
  photoUrl?: string | null;
}

export type UpdateEmployeeBodyStatus =
  (typeof UpdateEmployeeBodyStatus)[keyof typeof UpdateEmployeeBodyStatus];

export const UpdateEmployeeBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateEmployeeBody {
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string | null;
  position?: string;
  departmentId?: number | null;
  hireDate?: string;
  salary?: number;
  status?: UpdateEmployeeBodyStatus;
  address?: string | null;
  photoUrl?: string | null;
}

export type PayrollStatus = (typeof PayrollStatus)[keyof typeof PayrollStatus];

export const PayrollStatus = {
  pending: "pending",
  processed: "processed",
  paid: "paid",
} as const;

export interface Payroll {
  id: number;
  employeeId: number;
  employeeName?: string | null;
  month: number;
  year: number;
  basicSalary: number;
  bonuses: number;
  deductions: number;
  netSalary: number;
  status: PayrollStatus;
  processedAt?: string | null;
  createdAt: string;
}

export type CreatePayrollBodyStatus =
  (typeof CreatePayrollBodyStatus)[keyof typeof CreatePayrollBodyStatus];

export const CreatePayrollBodyStatus = {
  pending: "pending",
  processed: "processed",
  paid: "paid",
} as const;

export interface CreatePayrollBody {
  employeeId: number;
  month: number;
  year: number;
  basicSalary: number;
  bonuses: number;
  deductions: number;
  status: CreatePayrollBodyStatus;
}

export type LeaveRequestLeaveType =
  (typeof LeaveRequestLeaveType)[keyof typeof LeaveRequestLeaveType];

export const LeaveRequestLeaveType = {
  annual: "annual",
  sick: "sick",
  maternity: "maternity",
  paternity: "paternity",
  unpaid: "unpaid",
} as const;

export type LeaveRequestStatus =
  (typeof LeaveRequestStatus)[keyof typeof LeaveRequestStatus];

export const LeaveRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface LeaveRequest {
  id: number;
  employeeId: number;
  employeeName?: string | null;
  leaveType: LeaveRequestLeaveType;
  startDate: string;
  endDate: string;
  reason?: string | null;
  status: LeaveRequestStatus;
  reviewedBy?: number | null;
  reviewNote?: string | null;
  createdAt: string;
}

export type CreateLeaveBodyLeaveType =
  (typeof CreateLeaveBodyLeaveType)[keyof typeof CreateLeaveBodyLeaveType];

export const CreateLeaveBodyLeaveType = {
  annual: "annual",
  sick: "sick",
  maternity: "maternity",
  paternity: "paternity",
  unpaid: "unpaid",
} as const;

export interface CreateLeaveBody {
  employeeId: number;
  leaveType: CreateLeaveBodyLeaveType;
  startDate: string;
  endDate: string;
  reason?: string | null;
}

export type UpdateLeaveBodyStatus =
  (typeof UpdateLeaveBodyStatus)[keyof typeof UpdateLeaveBodyStatus];

export const UpdateLeaveBodyStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateLeaveBody {
  status: UpdateLeaveBodyStatus;
  reviewNote?: string | null;
}

export interface DashboardSummary {
  totalEmployees: number;
  activeEmployees: number;
  totalDepartments: number;
  totalPayrollThisMonth: number;
  pendingLeaves: number;
  newHiresThisMonth: number;
}

export interface PayrollByDepartment {
  departmentName: string;
  totalSalary: number;
  employeeCount: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  actor?: string | null;
}

export type ListEmployeesParams = {
  search?: string;
  departmentId?: number;
  status?: ListEmployeesStatus;
};

export type ListEmployeesStatus =
  (typeof ListEmployeesStatus)[keyof typeof ListEmployeesStatus];

export const ListEmployeesStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export type UploadEmployeePhotoBody = {
  photoUrl: string;
};

export type ListPayrollParams = {
  employeeId?: number;
  month?: number;
  year?: number;
};

export type ListLeavesParams = {
  employeeId?: number;
  status?: ListLeavesStatus;
};

export type ListLeavesStatus =
  (typeof ListLeavesStatus)[keyof typeof ListLeavesStatus];

export const ListLeavesStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;