# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR & Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from "./employeeStatus";

export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  position: string;
  departmentId?: number | null;
  departmentName?: string | null;
  hireDate: string;
  salary: number;
  status: EmployeeStatus;
  photoUrl?: string | null;
  address?: string | null;
  createdAt: string;
}