# lib/api-zod/src/generated/types/updateEmployeeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR & Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEmployeeBodyStatus } from "./updateEmployeeBodyStatus";

export interface UpdateEmployeeBody {
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string | null;
  position?: string;
  departmentId?: number | null;
  hireDate?: string;
  salary?: number;
  status?: UpdateEmployeeBodyStatus;
  address?: string | null;
  photoUrl?: string | null;
}